Changed Logic to accept case-insensitive commands. Cheers!

+++ src/main/java/seedu/address/logic/parser/AddressBookParser.java
@@ -46,45 +46,42 @@ public class AddressBookParser {
 
         final String commandWord = matcher.group("commandWord");
         final String arguments = matcher.group("arguments");
-        switch (commandWord) {
-
-        case AddCommand.COMMAND_WORD:
+        if (commandWord.equalsIgnoreCase(AddCommand.COMMAND_WORDVAR_1) || commandWord.equalsIgnoreCase(AddCommand.COMMAND_WORDVAR_2)) {
             return new AddCommandParser().parse(arguments);
 
-        case EditCommand.COMMAND_WORD:
+        } else if (commandWord.equalsIgnoreCase(EditCommand.COMMAND_WORDVAR_1) || commandWord.equalsIgnoreCase(EditCommand.COMMAND_WORDVAR_2)) {
             return new EditCommandParser().parse(arguments);
 
-        case SelectCommand.COMMAND_WORD:
+        } else if (commandWord.equalsIgnoreCase(SelectCommand.COMMAND_WORDVAR_1) || commandWord.equalsIgnoreCase(SelectCommand.COMMAND_WORDVAR_2)) {
             return new SelectCommandParser().parse(arguments);
 
-        case DeleteCommand.COMMAND_WORD:
+        } else if (commandWord.equalsIgnoreCase(DeleteCommand.COMMAND_WORDVAR_1) || commandWord.equalsIgnoreCase(DeleteCommand.COMMAND_WORDVAR_2)) {
             return new DeleteCommandParser().parse(arguments);
 
-        case ClearCommand.COMMAND_WORD:
+        } else if (commandWord.equalsIgnoreCase(ClearCommand.COMMAND_WORDVAR_1) || commandWord.equalsIgnoreCase(ClearCommand.COMMAND_WORDVAR_2)) {
             return new ClearCommand();
-
-        case FindCommand.COMMAND_WORD:
+        } else if (commandWord.equalsIgnoreCase(FindCommand.COMMAND_WORDVAR_1) || commandWord.equalsIgnoreCase(FindCommand.COMMAND_WORDVAR_2)) {
             return new FindCommandParser().parse(arguments);
 
-        case ListCommand.COMMAND_WORD:
+        } else if (commandWord.equalsIgnoreCase(ListCommand.COMMAND_WORDVAR_1) || commandWord.equalsIgnoreCase(ListCommand.COMMAND_WORDVAR_2)) {
             return new ListCommand();
 
-        case HistoryCommand.COMMAND_WORD:
+        } else if (commandWord.equalsIgnoreCase(HistoryCommand.COMMAND_WORDVAR_1) || commandWord.equalsIgnoreCase(HistoryCommand.COMMAND_WORDVAR_2)) {
             return new HistoryCommand();
 
-        case ExitCommand.COMMAND_WORD:
+        } else if (commandWord.equalsIgnoreCase(ExitCommand.COMMAND_WORD)) {
             return new ExitCommand();
 
-        case HelpCommand.COMMAND_WORD:
+        } else if (commandWord.equalsIgnoreCase(HelpCommand.COMMAND_WORD)) {
             return new HelpCommand();
 
-        case UndoCommand.COMMAND_WORD:
+        } else if (commandWord.equalsIgnoreCase(UndoCommand.COMMAND_WORDVAR_1) || commandWord.equalsIgnoreCase(UndoCommand.COMMAND_WORDVAR_2)) {
             return new UndoCommand();
 
-        case RedoCommand.COMMAND_WORD:
+        } else if (commandWord.equalsIgnoreCase(RedoCommand.COMMAND_WORDVAR_1) || commandWord.equalsIgnoreCase(RedoCommand.COMMAND_WORDVAR_2)) {
             return new RedoCommand();
 
-        default:
+        } else {
             throw new ParseException(MESSAGE_UNKNOWN_COMMAND);
         }
     }
